+++ src/components/button/base-button/BaseButton.tsx
@@ -1,30 +1,27 @@
 import React, { HTMLProps, ReactNode } from "react";
 import classNames from "classnames";
 import DataAttributes from "../../../types/DataAttributes";
-import { LoaderSize, Size } from "../../../types/Size";
-import { LoaderColor } from "../../../types/LoaderColors";
 import { Loader } from "../../loader/Loader";
-import { Shape } from "../../../types/Shape";
 
 export interface BaseButtonProps
 	extends Omit<HTMLProps<HTMLButtonElement>, "type" | "size">,
 		DataAttributes {
 	children?: ReactNode;
 	variant?: "filled" | "outline" | "link";
-	size?: Size;
+	size?: "small" | "medium" | "large";
 	loading?: boolean;
 	disabled?: boolean;
 	label?: string;
 	type?: "button" | "submit";
 	additionalClasses?: string[];
-	shape?: Shape;
+	shape?: "default" | "rounded" | "square";
 }
 
 const BaseButton = ({
 	children,
 	variant = "filled",
-	size = Size.MEDIUM,
-	color = LoaderColor.PRIMARY,
+	size = "medium",
+	color = "primary",
 	loading,
 	disabled,
 	className,
@@ -32,7 +29,7 @@ const BaseButton = ({
 	label = "button",
 	dataAttributes,
 	additionalClasses = [],
-	shape = Shape.DEFAULT,
+	shape = "default",
 	...props
 }: BaseButtonProps) => {
 	const classes = classNames(
@@ -43,12 +40,12 @@ const BaseButton = ({
 			"shadow-none focus:ring-2 border-2 border-gray-200 dark:border-gray-500 hover:border-primary-500 active:border-primary-400 dark:hover:border-gray-400 dark:active:border-gray-300 bg-white dark:bg-transparent text-primary-500 dark:text-gray-400",
 		variant === "link" &&
 			"shadow-none text-primary-500 hover:text-primary-400 active:text-primary-600",
-		size === Size.SMALL && "text-xs px-2 h-7",
-		size === Size.MEDIUM && "text-sm px-4 h-9",
-		size === Size.LARGE && "text-base px-6 h-11",
-		shape === Shape.DEFAULT && "rounded-md",
-		shape === Shape.ROUNDED && "rounded-full",
-		shape === Shape.SQUARE && "rounded-none",
+		size === "small" && "text-xs px-2 h-7",
+		size === "medium" && "text-sm px-4 h-9",
+		size === "large" && "text-base px-6 h-11",
+		shape === "default" && "rounded-md",
+		shape === "rounded" && "rounded-full",
+		shape === "square" && "rounded-none",
 		disabled && "opacity-50 cursor-not-allowed",
 		...additionalClasses,
 		className
@@ -63,11 +60,7 @@ const BaseButton = ({
 			{...props}
 			aria-label={label}
 		>
-			{loading ? (
-				<Loader size={LoaderSize.EXTRASMALL} color={LoaderColor.CURRENT} />
-			) : (
-				children
-			)}
+			{loading ? <Loader size="xs" color="currentColor" /> : children}
 		</button>
 	);
 };