+++ src/components/field/input/Input.tsx
@@ -1,5 +1,5 @@
 import { FieldBase } from "../field-base/FieldBase";
-import { ChangeEvent, forwardRef, ReactNode, useState } from "react";
+import { ChangeEvent, forwardRef, ReactNode, useId, useState } from "react";
 import classNames from "classnames";
 import { InputTypes } from "../../../types/InputTypes";
 import { Affix } from "../affix/Affix";
@@ -37,7 +37,6 @@ interface InputProps extends DataAttributes {
 	autocomplete?: string;
 	size?: Size;
 	shape?: Shape;
-	formField?: boolean;
 }
 
 export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
@@ -63,80 +62,74 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
 		dataAttributes,
 		size = Size.MEDIUM,
 		shape = Shape.DEFAULT,
-		formField = true,
 	},
 	ref
 ) {
-	const classes = classNames("flex items-center gap-2", className);
-	const wrapperClasses = classNames("input-wrapper", {
-		"form-field": formField,
-	});
-
+	const uniqueId = useId();
+	const classes = classNames("flex flex-col gap-2", className);
 	const [withinFocus, setWithinFocus] = useState(false);
 
 	return (
-		<div className={wrapperClasses}>
-			<div className={classes} {...dataAttributes}>
-				{label && (
-					<Label id={id} required={required}>
-						{label}
-					</Label>
-				)}
-				<FieldBase
+		<div className={classes} {...dataAttributes}>
+			{label && (
+				<Label id={uniqueId} required={required}>
+					{label}
+				</Label>
+			)}
+			<FieldBase
+				disabled={disabled}
+				readonly={readonly}
+				error={error}
+				label={label}
+				shape={shape}
+				size={size}
+				className="px-1!"
+				activated={withinFocus}
+			>
+				{prefix}
+				<input
+					id={uniqueId}
+					ref={ref}
+					className="border-none bg-transparent! outline-none w-full focus:ring-0 focus:ring-offset-0 h-8"
+					value={value}
+					type={inputType}
+					onChange={
+						onChange || onChangeEvent
+							? (e) => {
+									if (onChange) onChange(e.target.value);
+									if (onChangeEvent) onChangeEvent(e);
+							  }
+							: undefined
+					}
+					placeholder={placeholder}
+					readOnly={readonly}
 					disabled={disabled}
-					readonly={readonly}
-					error={error}
-					label={label}
-					shape={shape}
-					size={size}
-					className="px-1!"
-					activated={withinFocus}
-				>
-					{prefix}
-					<input
-						id={id}
-						ref={ref}
-						className="border-none bg-transparent! outline-none w-full focus:ring-0 focus:ring-offset-0"
-						value={value}
-						type={inputType}
-						onChange={
-							onChange || onChangeEvent
-								? (e) => {
-										if (onChange) onChange(e.target.value);
-										if (onChangeEvent) onChangeEvent(e);
-								  }
-								: undefined
-						}
-						placeholder={placeholder}
-						readOnly={readonly}
-						disabled={disabled}
-						autoComplete={autocomplete}
-						onFocus={() => setWithinFocus(true)}
-						onBlur={(e) => {
-							setWithinFocus(false);
-							onBlur?.(e as ChangeEvent<HTMLInputElement>);
+					autoComplete={autocomplete}
+					onFocus={() => setWithinFocus(true)}
+					onBlur={(e) => {
+						setWithinFocus(false);
+						onBlur?.(e as ChangeEvent<HTMLInputElement>);
+					}}
+				/>
+				{clearable && (
+					<Affix
+						className={{
+							"cursor-pointer": true,
+							visible: value !== "",
 						}}
-					/>
-					{clearable && (
-						<Affix
-							className={{
-								"cursor-pointer": true,
-								visible: value !== "",
-							}}
-							onClick={() =>
-								onChangeEvent &&
-								onChangeEvent({
-									target: { value: "" },
-								} as ChangeEvent<HTMLInputElement>)
-							}
-						>
-							<XMarkIcon className="w-4 h-4" data-testid="clear-button" />
-						</Affix>
-					)}
-					{suffix}
-				</FieldBase>
-				{errorMsg && error && <ErrorMessage>{errorMsg}</ErrorMessage>}
-			</div>
+						onClick={() =>
+							onChangeEvent &&
+							onChangeEvent({
+								target: { value: "" },
+							} as ChangeEvent<HTMLInputElement>)
+						}
+					>
+						<XMarkIcon className="w-4 h-4" data-testid="clear-button" />
+					</Affix>
+				)}
+				{suffix}
+			</FieldBase>
+			{errorMsg && error && <ErrorMessage>{errorMsg}</ErrorMessage>}
 		</div>
 	);
 });