+++ src/types/LoaderColors.ts
@@ -1,7 +1,0 @@
-export enum LoaderColor {
-	PRIMARY = "primary",
-	WARNING = "warning",
-	DANGER = "danger",
-	SUCCESS = "success",
-	CURRENT = "currentColor",
-}